requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
numpy>=1.24.0
mediapipe>=0.10.0
twilio>=8.5.0
python-dotenv>=1.0.0
playsound>=1.3.0
pandas>=2.0.0
scikit-learn>=1.2.0
moviepy>=1.0.3
pytorchvideo>=0.1.5
av>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
iopath>=0.1.10
fvcore>=0.1.5 